# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  type: string
  description: string
  /** @nullable */
  userId?: number | null
  createdAt: string
}