# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomQuality } from './roomQuality'
import type { User } from './user'

export interface Room {
  id: string
  name: string
  inviteCode: string
  ownerId: number
  memberCount: number
  members?: User[]
  quality?: RoomQuality
  isActive?: boolean
  createdAt: string
}