# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AetherPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List current user's rooms
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomInput'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /rooms/{roomId}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room by invite code
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomInput'
      responses:
        '200':
          description: Joined room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '403':
          description: Invalid invite code
        '404':
          description: Room not found
  /rooms/join-by-code:
    post:
      operationId: joinRoomByCode
      tags: [rooms]
      summary: Join a room using only an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomInput'
      responses:
        '200':
          description: Joined room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '403':
          description: Invalid invite code
        '404':
          description: Room not found
  /rooms/{roomId}/leave:
    post:
      operationId: leaveRoom
      tags: [rooms]
      summary: Leave a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Left room
...
  /rooms/{roomId}/activity:
    get:
      operationId: getRoomActivity
      tags: [rooms]
      summary: Get room activity summary (member count, recent events)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomActivity'
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ['string', 'null']
        bannerUrl:
          type: ['string', 'null']
        bio:
          type: ['string', 'null']
        pronouns:
          type: ['string', 'null']
        website:
          type: ['string', 'null']
        location:
          type: ['string', 'null']
        status:
          type: string
          enum: [online, away, busy, offline]
        customStatus:
          type: ['string', 'null']
        accentColor:
          type: ['string', 'null']
        primaryColor:
          type: ['string', 'null']
        displayNameStyle:
          type: ['string', 'null']
        profileGradient:
          type: ['string', 'null']
        avatarFrame:
          type: ['string', 'null']
        profileTheme:
          type: ['string', 'null']
        customTheme:
          type: ['string', 'null']
        preferredTheme:
          type: ['string', 'null']
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Room:
      type: object
      required: [id, name, inviteCode, ownerId, memberCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        inviteCode:
          type: string
        ownerId:
          type: integer
        memberCount:
          type: integer
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        quality:
          type: string
          enum: [360p, 480p, 720p, 1080p, 1440p]
        isActive:
          type: boolean
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        quality:
          type: string
          enum: [360p, 480p, 720p, 1080p, 1440p]
    JoinRoomInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
...
    RoomActivity:
      type: object
      required: [roomId, memberCount, messageCount, events]
      properties:
        roomId:
          type: string
        memberCount:
          type: integer
        messageCount:
          type: integer
        events:
          type: array
          items:
            $ref: '#/components/schemas/ActivityEvent'
    ActivityEvent:
      type: object
      required: [type, description, createdAt]
      properties:
        type:
          type: string
        description:
          type: string
        userId:
          type: ['integer', 'null']
        createdAt:
          type: string
...